Remove CompositeMetadata, make Entry child of flyweight
Entry is now the only high-level abstraction, will need a way to
decode into a List<Entry>.
