+++ rsocket-core/src/main/java/io/rsocket/metadata/CompositeMetadataFlyweight.java
@@ -192,8 +192,8 @@ public class CompositeMetadataFlyweight {
     // see #encodeMetadataHeader(ByteBufAllocator, byte, int)
     public static void encodeAndAddMetadata(CompositeByteBuf compositeMetaData, ByteBufAllocator allocator, WellKnownMimeType knownMimeType, ByteBuf metadata) {
         compositeMetaData.addComponents(true,
-                encodeMetadataHeader(allocator, knownMimeType.getIdentifier(), metadata.readableBytes()),
-                metadata);
+                encodeMetadataHeader(allocator, knownMimeType.getIdentifier(),
+                        metadata.readableBytes()), metadata);
     }
 
     /**
@@ -207,8 +207,239 @@ public class CompositeMetadataFlyweight {
     // see #encodeMetadataHeader(ByteBufAllocator, byte, int)
     static void encodeAndAddMetadata(CompositeByteBuf compositeMetaData, ByteBufAllocator allocator, byte unknownCompressedMimeType, ByteBuf metadata) {
         compositeMetaData.addComponents(true,
-                encodeMetadataHeader(allocator, unknownCompressedMimeType, metadata.readableBytes()),
-                metadata);
+                encodeMetadataHeader(allocator, unknownCompressedMimeType,
+                        metadata.readableBytes()), metadata);
     }
 
+    //=== ENTRY ===
+
+    /**
+     * An entry in a Composite Metadata, which exposes the {@link #getMimeType() mime type} and {@link ByteBuf}
+     * {@link #getMetadata() content} of the metadata entry.
+     */
+    public interface Entry {
+
+        /**
+         * Create an {@link Entry} from a {@link WellKnownMimeType}. This will be encoded in a compressed format that
+         * uses the {@link WellKnownMimeType#getIdentifier() mime identifier}.
+         *
+         * @param mimeType the {@link WellKnownMimeType} to use for the entry
+         * @param metadataContentBuffer the content {@link ByteBuf} to use for the entry
+         * @return the new entry
+         */
+        public static Entry wellKnownMime(WellKnownMimeType mimeType, ByteBuf metadataContentBuffer) {
+            return new CompressedTypeEntry(mimeType, metadataContentBuffer);
+        }
+
+        /**
+         * Create an {@link Entry} from a custom mime type represented as an US-ASCII only {@link String}.
+         * The whole literal mime type will thus be encoded.
+         *
+         * @param mimeType the custom mime type {@link String}
+         * @param metadataContentBuffer the content {@link ByteBuf} to use for the entry
+         * @return the new entry
+         */
+        public static Entry customMime(String mimeType, ByteBuf metadataContentBuffer) {
+            return new CustomTypeEntry(mimeType, metadataContentBuffer);
+        }
+
+        /**
+         * Create an {@link Entry} from an unrecognized yet valid "well-known" mime type, ie. a {@code byte} that would map
+         * to {@link WellKnownMimeType#UNKNOWN_RESERVED_MIME_TYPE}. Prefer using {@link #wellKnownMime(WellKnownMimeType, ByteBuf)}
+         * if the mime code is recognizable by this client.
+         * <p>
+         * This case would usually be encountered when decoding a composite metadata entry from a remote that uses a more recent
+         * version of the {@link WellKnownMimeType} extension, and this method can be useful to create an unprocessed entry
+         * in such a case, ensuring no information is lost when forwarding frames.
+         *
+         * @param mimeCode the reserved but unrecognized compressed mime type {@code byte}
+         * @param metadataContentBuffer the content {@link ByteBuf} to use for the entry
+         * @return the new entry
+         * @see #wellKnownMime(WellKnownMimeType, ByteBuf)
+         */
+        public static Entry rawCompressedMime(byte mimeCode, ByteBuf metadataContentBuffer) {
+            return new UnknownCompressedTypeEntry(mimeCode, metadataContentBuffer);
+        }
+
+        /**
+         * Incrementally decode the next metadata entry from a {@link ByteBuf} into an {@link Entry}.
+         * This is only possible on frame types used to initiate
+         * interactions, if the SETUP metadata mime type was {@link WellKnownMimeType#MESSAGE_RSOCKET_COMPOSITE_METADATA}.
+         * <p>
+         * Each entry {@link ByteBuf} is a  {@link ByteBuf#readSlice(int) slice} of the original buffer that can also be
+         * {@link ByteBuf#readRetainedSlice(int) retained} if needed.
+         *
+         * @param buffer the buffer to decode
+         * @param retainMetadataSlices should each slide be retained when read from the original buffer?
+         * @return the decoded {@link Entry}
+         */
+        static Entry decodeEntry(ByteBuf buffer, boolean retainMetadataSlices) {
+            ByteBuf[] entry = decodeMimeAndContentBuffers(buffer, retainMetadataSlices);
+            if (entry == METADATA_MALFORMED) {
+                throw new IllegalArgumentException("composite metadata entry buffer is too short to contain proper entry");
+            }
+            if (entry == METADATA_BUFFERS_DONE) {
+                return null;
+            }
+
+            ByteBuf encodedHeader = entry[0];
+            ByteBuf metadataContent = entry[1];
+
+
+            //the flyweight already validated the size of the buffer,
+            //this is only to distinguish id vs custom type
+            if (encodedHeader.readableBytes() == 1) {
+                //id
+                byte id = decodeMimeIdFromMimeBuffer(encodedHeader);
+                WellKnownMimeType wkn = WellKnownMimeType.fromId(id);
+                if (wkn == WellKnownMimeType.UNPARSEABLE_MIME_TYPE) {
+                    //should not happen due to flyweight's decodeEntry own guard
+                    throw new IllegalStateException("composite metadata entry parsing failed on compressed mime id " + id);
+                }
+                if (wkn == WellKnownMimeType.UNKNOWN_RESERVED_MIME_TYPE) {
+                    return new UnknownCompressedTypeEntry(id, metadataContent);
+                }
+                return new CompressedTypeEntry(wkn, metadataContent);
+            }
+            else {
+                CharSequence customMimeCharSequence = decodeMimeTypeFromMimeBuffer(encodedHeader);
+                if (customMimeCharSequence == null) {
+                    //should not happen due to flyweight's decodeEntry own guard
+                    throw new IllegalArgumentException("composite metadata entry parsing failed on custom type");
+                }
+                return new CustomTypeEntry(customMimeCharSequence.toString(), metadataContent);
+            }
+        }
+
+        /**
+         * A <i>passthrough</i> entry is one for which the {@link #getMimeType()} could not be decoded.
+         * This is usually because it was compressed on the wire, but using an id that is still just "reserved for
+         * future use" in this implementation.
+         * <p>
+         * Still, another actor on the network might be able to interpret such an entry, which should thus be
+         * re-encoded as it was when forwarding the frame.
+         * <p>
+         * The {@link #getMetadata()} exposes the raw content buffer of the entry (as any other entry).
+         *
+         * @return true if this entry should be ignored but passed through as is during re-encoding
+         */
+        default boolean isPassthrough() {
+            return false;
+        }
+
+        /**
+         * @return the mime type for this entry
+         */
+        @Nullable
+        String getMimeType();
+
+        /**
+         * @return the metadata content of this entry
+         */
+        ByteBuf getMetadata();
+
+        /**
+         * Encode this {@link Entry} into a {@link CompositeByteBuf} representing a composite metadata.
+         * This buffer may already hold components for previous {@link Entry entries}.
+         *
+         * @param compositeByteBuf the {@link CompositeByteBuf} to hold the components of the whole composite metadata
+         * @param byteBufAllocator the {@link ByteBufAllocator} to use to allocate new buffers as needed
+         */
+        default void encodeInto(CompositeByteBuf compositeByteBuf, ByteBufAllocator byteBufAllocator) {
+            if (this instanceof CompressedTypeEntry) {
+                CompressedTypeEntry cte = (CompressedTypeEntry) this;
+                encodeAndAddMetadata(compositeByteBuf, byteBufAllocator,
+                        cte.mimeType, cte.content);
+            }
+            else if (this instanceof UnknownCompressedTypeEntry) {
+                UnknownCompressedTypeEntry ucte = (UnknownCompressedTypeEntry) this;
+                encodeAndAddMetadata(compositeByteBuf, byteBufAllocator,
+                        (byte) ucte.identifier, ucte.content);
+            }
+            else {
+                encodeAndAddMetadata(compositeByteBuf, byteBufAllocator,
+                        getMimeType(), getMetadata());
+            }
+        }
+    }
+
+    static final class CustomTypeEntry implements Entry {
+
+        private final String mimeType;
+        private final ByteBuf content;
+
+        CustomTypeEntry(String mimeType, ByteBuf content) {
+            this.mimeType = mimeType;
+            this.content = content;
+        }
+
+        @Override
+        public String getMimeType() {
+            return this.mimeType;
+        }
+
+        @Override
+        public ByteBuf getMetadata() {
+            return this.content;
+        }
+    }
+
+    static final class CompressedTypeEntry implements Entry {
+
+        private final WellKnownMimeType mimeType;
+        private final ByteBuf content;
+
+        CompressedTypeEntry(WellKnownMimeType mimeType, ByteBuf content) {
+            this.mimeType = mimeType;
+            this.content = content;
+        }
+
+        @Override
+        public String getMimeType() {
+            return this.mimeType.getMime();
+        }
+
+        @Override
+        public ByteBuf getMetadata() {
+            return this.content;
+        }
+    }
+
+    /**
+     * A {@link Entry#isPassthrough() pass-through} {@link Entry} that encapsulates a
+     * compressed-mime metadata that couldn't be recognized.
+     */
+    static final class UnknownCompressedTypeEntry implements Entry {
+
+        private final byte identifier;
+        private final ByteBuf content;
+
+        UnknownCompressedTypeEntry(byte identifier, ByteBuf content) {
+            this.identifier = identifier;
+            this.content = content;
+        }
+
+        @Override
+        public boolean isPassthrough() {
+            return true;
+        }
+
+        @Override
+        public String getMimeType() {
+            return WellKnownMimeType.UNKNOWN_RESERVED_MIME_TYPE.getMime();
+        }
+
+        /**
+         * Return the compressed identifier that was used in the original decoded metadata, but couldn't be
+         * decoded because this implementation only knows this as "reserved for future use". The original
+         */
+        public byte getUnknownReservedId() {
+            return this.identifier;
+        }
+
+        @Override
+        public ByteBuf getMetadata() {
+            return this.content;
+        }
+    }
 }
